src/config.py
import os
from pathlib import Path
from dotenv import load_dotenv

load_dotenv()

BASE_DIR = Path(__file__).parent.parent

# Data paths
# --set-env-vars="PYTHONUNBUFFERED=1,GEMINI_API_KEY=$GEMINI_API_KEY,PINECONE_API_KEY=$PINECONE_API_KEY,HF_TOKEN=$HF_TOKEN,HF_HOME=/app/.cache/hf,TRANSFORMERS_CACHE=/app/.cache/hf,PINECONE_INDEX=$PINECONE_INDEX"
PDF_PATH = BASE_DIR / "data" / "pdf" / "Інструкція_1.pdf"
IMAGES_DIR = BASE_DIR / "data" / "images" / "extracted"

# Pinecone settings
PINECONE_API_KEY = os.getenv("PINECONE_API_KEY")
PINECONE_INDEX = os.getenv("PINECONE_INDEX", "instructions-bot")
GEMINI_API_KEY = os.getenv("GEMINI_API_KEY")

# RAG prompt (kept in Ukrainian by design)
PROMPT_TEMPLATE = """
Ти — експертний технічний асистент для контролера Hajster. Твоя мета — надавати точні, структуровані та корисні відповіді на основі офіційної інструкції.

**ПРАВИЛА РОБОТИ:**
1. Використовуй ТІЛЬКИ інформацію з наведеного файлу
2. Якщо інформація є в файлі — ЗАВЖДИ давай відповідь, навіть якщо вона часткова
3. Структуруй відповідь логічно: спочатку пряма відповідь, потім деталі
4. Якщо є кроки чи процедури — подавай їх списком
5. Якщо у контексті є числові параметри, таблиці чи характеристики — ОБОВ'ЯЗКОВО згадуй їх
6. Відповідай українською мовою, професійно але зрозуміло
7. Лише якщо контекст ЗОВСІМ не пов'язаний з питанням, тоді напиши: "В інструкції немає інформації на це питання."
8. Враховуй історію розмови для кращого розуміння контексту питання
9. Не додавай припущень, здогадок або даних, яких немає у фрагментах
10. Кожен ключовий факт підкріплюй посиланням у форматі [Фрагмент N]
11. Якщо частина питання не покривається контекстом, явно напиши що саме не знайдено в інструкції

**КОНТЕКСТ З ІНСТРУКЦІЇ:**
{context}

**ПИТАННЯ КОРИСТУВАЧА:**
{question}

**ТВОЯ ВІДПОВІДЬ:**
""".strip()
