requirements.txt
streamlit
sentence-transformers==2.7.0
sentencepiece
langchain
langchain-classic
langchain-core
langchain-text-splitters
langchain-google-genai
pinecone
python-dotenv
PyMuPDF
docling
transformers==4.55.2
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
Flask
Flask-Cors
gunicorn
rapidocr-onnxruntime
